Moved UitpasNumberValidator and other Uitpas functions to the Uitpas service
