+++ backend/app/api/src/boot.ts
@@ -20,7 +20,7 @@ import { PlatformMembershipService } from './services/PlatformMembershipService'
 import { UniqueUserService } from './services/UniqueUserService';
 import { QueueHandler } from '@stamhoofd/queues';
 import { SimpleError } from '@simonbackx/simple-errors';
-import { UitpasNumberService } from './services/UitpasNumbersService';
+import { UitpasService } from './services/UitpasService';
 
 process.on('unhandledRejection', (error: Error) => {
     console.error('unhandledRejection');
@@ -230,7 +230,7 @@ const start = async () => {
     PlatformMembershipService.listen();
     DocumentService.listen();
     SetupStepUpdater.listen();
-    UitpasNumberService.listen();
+    UitpasService.listen();
 
     startCrons();
     seeds().catch(console.error);
